x86/boot: Convert early_serial_base to unsigned long
As a preparatory of adding flexible serial I/O accessors, convert
early_serial_base to unsigned long to cover all possible bus addresses
on the system.

No functional change intended.

Link: https://lore.kernel.org/r/20250502123145.4066635-2-andriy.shevchenko@linux.intel.com

+++ arch/x86/boot/tty.c
@@ -13,7 +13,7 @@
 
 #include "boot.h"
 
-int early_serial_base;
+unsigned long early_serial_base;
 
 #define XMTRDY          0x20
